# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise Ticketing, Project Management & Workflow System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── TICKETS ──────────────────────────────────────────────────────────────────
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: myProjects
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /projects/{id}/progress:
    patch:
      operationId: updateProjectProgress
      tags: [projects]
      summary: Update project progress
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectProgressUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/collaborators:
    post:
      operationId: addProjectCollaborator
      tags: [projects]
      summary: Add collaborator to project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollaboratorInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/collaborators/{userId}:
    delete:
      operationId: removeProjectCollaborator
      tags: [projects]
      summary: Remove collaborator from project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # ── TO-DO ────────────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /calendar/events:
    get:
      operationId: listCalendarEvents
      tags: [calendar]
      summary: List calendar events
      parameters:
        - name: fromDate
          in: query
          schema: { type: string }
        - name: toDate
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createCalendarEvent
      tags: [calendar]
      summary: Create calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /calendar/events/{id}:
    delete:
      operationId: deleteCalendarEvent
      tags: [calendar]
      summary: Delete calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── TIMESHEETS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, type, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        description: { type: ["string", "null"] }
        createdAt: { type: string }
    CategoryInput:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type: { type: string }
        description: { type: string }

    # ── TICKET SCHEMAS ──────────────────────────────────────────────────────
...
    Project:
      type: object
      required: [id, projectNo, title, status, priority, createdAt]
      properties:
        id: { type: integer }
        projectNo: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        category: { type: ["string", "null"] }
        type: { type: ["string", "null"] }
        progress: { type: integer }
        ownerId: { type: ["integer", "null"] }
        ownerName: { type: ["string", "null"] }
        processOwnerId: { type: ["integer", "null"] }
        processOwnerName: { type: ["string", "null"] }
        collaborators:
          type: array
          items:
            $ref: "#/components/schemas/ProjectCollaborator"
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        reviewFrequency: { type: ["string", "null"] }
        ticketCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProjectCollaborator:
      type: object
      required: [userId, name]
      properties:
        userId: { type: integer }
        name: { type: string }
        avatarUrl: { type: ["string", "null"] }
        role: { type: ["string", "null"] }
    ProjectInput:
      type: object
      required: [title, priority]
      properties:
        title: { type: string }
        description: { type: string }
        priority: { type: string }
        category: { type: string }
        type: { type: string }
        ownerId: { type: integer }
        processOwnerId: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
        reviewFrequency: { type: string }
    ProjectUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        status: { type: string }
        priority: { type: string }
        ownerId: { type: integer }
        processOwnerId: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
    ProjectProgressUpdate:
      type: object
      required: [progress]
      properties:
        progress: { type: integer }
        remarks: { type: string }
    CollaboratorInput:
      type: object
      required: [userId]
      properties:
        userId: { type: integer }
        role: { type: string }

    # ── TODO SCHEMAS ────────────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, type, message, isRead, createdAt]
      properties:
        id: { type: integer }
        type: { type: string }
        message: { type: string }
        entityType: { type: ["string", "null"] }
        entityId: { type: ["integer", "null"] }
        entityRef: { type: ["string", "null"] }
        isRead: { type: boolean }
        createdAt: { type: string }

    # ── CALENDAR SCHEMAS ────────────────────────────────────────────────────
    CalendarEvent:
      type: object
      required: [id, title, startDate, type, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        startDate: { type: string }
        endDate: { type: ["string", "null"] }
        type: { type: string }
        entityType: { type: ["string", "null"] }
        entityId: { type: ["integer", "null"] }
        meetingLink: { type: ["string", "null"] }
        createdAt: { type: string }
    CalendarEventInput:
      type: object
      required: [title, startDate, type]
      properties:
        title: { type: string }
        description: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        type: { type: string }
        meetingLink: { type: string }

    # ── TIMESHEET SCHEMAS ───────────────────────────────────────────────────
...